requirements/base.txt
# Versions are left out on purpose in order for you to use the latest ones.
# After the first "pip install -r", just run "pip freeze" and add the version
# to each package in each requirements/*.txt.

alabaster==0.7.6
Babel==2.1.1
binaryornot==0.4.0
chardet==2.3.0
click==5.1
cookiecutter==1.3.0
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.6
django-braces==1.8.1
django-model-utils==2.3.1
docutils==0.12
funcsigs==0.4
future==0.15.2
gunicorn==19.3.0
Jinja2==2.8
MarkupSafe==0.23
mock==1.3.0
model-mommy==1.2.5.1
pbr==1.8.1
Pillow==3.0.0
psycopg2==2.6.1
Pygments==2.0.2
pystache==0.5.4
pytz==2015.7
ruamel.base==1.0.0
ruamel.ordereddict==0.4.9
ruamel.yaml==0.10.12
six==1.10.0
snowballstemmer==1.2.0
Sphinx==1.3.1
sphinx-rtd-theme==0.1.9
static==1.1.1
static3==0.6.1
whichcraft==0.1.1